El tipo de dato de valor ahora es double mejorando el calculo de calificaciones

+++ models/actividad.js
@@ -17,7 +17,7 @@ module.exports = (sequelize, DataTypes) => {
     },
     nombre: DataTypes.STRING,
       descripcion: DataTypes.STRING,
-      valor: DataTypes.INTEGER,
+      valor: DataTypes.DOUBLE,
       tipo: DataTypes.STRING
   },
       {
